refactor: move http_client... to client_utils.py
The utils.py file has become large and disorganized.
Moving http_client(), HttpClientParams(),
and get_ssl_cert_config(), to client_utils.py would improve
organization and better align functionality
with the file's purpose.
